# gateway/openapi.yaml
openapi: 3.1.0
info:
  title: CareConnect API
  version: 1.0.0
  description: |
    Personal medical clipboard API with PHI isolation, QR consent, and Doctor Finder.

    **Security Model**: Zero-trust, idempotence required for mutations, TLS 1.3.
    **Compliance**: PHI confined to GCP Healthcare API; consent-gated reads.
    **Residency**: Canada (northamerica-northeast1).
  contact:
    name: CareConnect Platform Team
    email: platform@careconnect.health
servers:
  - url: https://api-dev.careconnect.health/v1
    description: Development (CA-dev)
  - url: http://localhost:8080/v1
    description: Local development
security:
  - BearerAuth: []
paths:
...
  /doctor-finder/subscribe:
    post:
      operationId: subscribeDoctorFinder
      summary: Opt-in to Doctor Finder concierge
      description: |
        Subscribe to receive warm intros for local providers.
        Uses compliant sources only (no scraping). Human-in-loop.
      tags:
        - Doctor Finder
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DoctorFinderSubscription'
      responses:
        '201':
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoctorFinderSubscriptionResult'
        '200':
          description: Idempotent replay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoctorFinderSubscriptionResult'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
  /doctor-finder/matches:
    get:
      operationId: getDoctorMatches
      summary: Get matched providers
      description: |
        Retrieve providers matched to subscriber's preferences.
        Returns non-PHI data only.
      tags:
        - Doctor Finder
      parameters:
        - name: subscriberId
          in: query
          required: true
          description: Subscription ID
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Matched providers
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DoctorMatch'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC JWT token with tenant/role claims
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: UUID v4 for idempotent mutations
      schema:
        type: string
        format: uuid
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: Request correlation ID for tracing
      schema:
        type: string
  responses:
...
    IdempotencyConflict:
      description: Idempotency-Key reused with different payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    DoctorFinderSubscription:
      type: object
      required:
        - region
        - language
      properties:
        region:
          type: string
          description: Province or city (e.g., AB, Calgary)
        language:
          type: string
          enum: [en, fr]
        specialty:
          type: string
          description: Desired specialty (e.g., family, cardiology)
        constraints:
          type: object
          description: Additional preferences (accepting patients, gender, etc.)
    DoctorFinderSubscriptionResult:
      allOf:
        - $ref: '#/components/schemas/DoctorFinderSubscription'
        - type: object
          required:
            - subscriberId
            - createdAt
          properties:
            subscriberId:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
    DoctorMatch:
      type: object
      required:
        - providerId
        - name
        - specialty
      properties:
        providerId:
          type: string
        name:
          type: string
        specialty:
          type: string
        location:
          type: object
          properties:
            city:
              type: string
            province:
              type: string
        acceptingPatients:
          type: boolean
        languages:
          type: array
          items:
            type: string
        sourceUrl:
          type: string
          description: Public directory URL
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        correlationId:
          type: string
          description: Request correlation ID